apps/IMAGE ID PRINT/ai-sidecar/requirements.txt
Flask==3.1.1
flask-cors==5.0.0
rembg==2.0.74
onnxruntime>=1.24.4
opencv-python-headless>=4.12.0.88
Pillow>=12.1.1,<13
